workaround kernel spidev warning
spidev is deprecated in newer kernels and provides a nasty error
message.  For base images provided by SR we include spidev for
easy testing by customers.  We can work around the errors by
using linux,spidev as the warning only checks for spidev, while
the modalias will still match properly with linux,spidev.

This is not recommended for production systems

Signed-off-by: Jon Nettleton <jon@solid-run.com>