re-import that patches from the main kernel source tree
This includes fixes for the HummingBoard Pulse with only a
Basler camera connected to the carrier. NXP's imx8-isp software
is very specific about how the mipi-csi, isp, and v4l entries
are setup, and having only the carrier CSI port enabled was
causing initialization failures. For now we will enabled both
the SOM and carrier ports to work around this.

Signed-off-by: Jon Nettleton <jon@solid-run.com>