arm-compute-library: Additional changes to fix build
As noted in the original recipe the arm-compute-library does not
support the build variables MAXLENGTH,PREFIX, or prefix that are
all injected by the default poky scons.bbclass. This should be
extended out to also the configure stage. This bbappend removes
the MAXLINELENGTH and then uses a custom do_configure() stage
that is taken from the scons.bbclass but removes the PREFIX and
prefix variables.

Signed-off-by: Jon Nettleton <jon@solid-run.com>