imx8dxl-sr-som: update gnss baudrate to 115200
Add a new setup command to execute before gpsd that reconfigures the
MIA-M10Q gps receiver for 115200 baudrate and issues soft reset on warm
boot.

This is required before starting gpsd, to ensure the module is in a sane
state as the gpsd ubx driver is not capable of cleanup and has buggy
behaviour for baudrate switching.